% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppFun}
\alias{CppFun}
\title{An Rcpp function for Exercise 9.8 (page 278, Statistical Computing with R).}
\usage{
CppFun(a, b, n, M)
}
\arguments{
\item{a}{A parameter the bivariate density.}

\item{b}{A parameter the bivariate density.}

\item{n}{A parameter the bivariate density.}

\item{M}{Length of the chain.}
}
\value{
A chain.
}
\description{
An Rcpp function for Exercise 9.8 (page 278, Statistical Computing with R).
}
