\name{StatComp21078-package}
\alias{StatComp21078-package}
\alias{StatComp21078}
\docType{package}
\title{
  Final homework of statistical computations
}
\description{
  It includes the homework of the whole semester and two functions of my research field.
}
\author{
StatComp21078
Maintainer: StatComp21078 <709476524@qq.com>
}
\keyword{ package }